WEBCMS-474 Improve Usability With Empty Selections
Disable all context sensitive folder actions if the current selection
is empty. Some of the actions act weirdly if triggered with no
documents, eg. the workflow managment form uses silently the parent
folder as a fallback. This is confusing and prevented by disabling the
access to these action in this case.